wip moisture precision

+++ measure.h
@@ -18,7 +18,7 @@ DHT dht(DHTPIN, DHTTYPE);
 struct measurement {
   unsigned long time;
   long weight; // 2,147,483,647 / 60,000 =  30000
-  long moisture;
+  float moisture;
   long temp;
   long humidity;
   long light;
